Renamed the extension class to follow the Behat shortcut convention
